+++ plydata/tests/test_cat_tools.py
@@ -1,6 +1,7 @@
 import pytest
 from plydata.cat_tools import (
-    cat_reorder2
+    cat_reorder2,
+    cat_shuffle,
 )
 
 
@@ -10,3 +11,10 @@ def test_reorder2():
     y = [1, 2, 3, 4, 5, 6]
     with pytest.raises(ValueError):
         cat_reorder2(c, x, y+[3])
+
+
+def test_shuffle():
+    c = list('abcde')
+
+    with pytest.raises(TypeError):
+        cat_shuffle(c, 'bad_random_state')